♻️([dogwood|eucalyptus]/3/[fun|wb]) move fun.py to docker_run_production
The specific FUN are a fine-tuning of the main settings defined in
docker_run_production.py. They should come after instead of being imported
on top of the docker_run_production.py file.

For example, this will allow us to derive the location where we store
certificates from the location where we store media files.
