+++ releases/eucalyptus/3/wb/CHANGELOG.md
@@ -9,6 +9,10 @@ release.
 
 ## [Unreleased]
 
+### Changed
+
+- Refactor FUN settings as fine tuning that comes after the main settings
+
 ## [eucalyptus.3-wb-1.3.1] - 2019-12-27
 
 ### Fixed
