+++ releases/eucalyptus/3/wb/config/cms/docker_run_production.py
@@ -23,7 +23,8 @@ from xmodule.modulestore.modulestore_settings import (
     update_module_store_settings,
 )
 
-from .fun import *
+from ..common import *
+from lms.envs.fun.utils import Configuration, prefer_fun_video
 
 # Load custom configuration parameters from yaml files
 config = Configuration(os.path.dirname(__file__))
@@ -742,10 +743,57 @@ LTI_XBLOCK_CONFIGURATIONS = config(
                 "ask_to_send_username": True,
                 "launch_target": "new_window",
             },
-        },
+        }
     ],
     formatter=json.loads,
 )
 LTI_XBLOCK_SECRETS = config(
     "LTI_XBLOCK_SECRETS", default={}, formatter=json.loads
 )
+
+################################ FUN stuff ################################
+
+# Fun-apps configuration
+INSTALLED_APPS += (
+    "fun",
+    "videoproviders",
+    "teachers",
+    "courses",
+    "haystack",
+    "universities",
+    "easy_thumbnails",
+    "ckeditor",
+    "raven.contrib.django.raven_compat",
+)
+
+ROOT_URLCONF = "fun.cms.urls"
+
+# ### THIRD-PARTY SETTINGS ###
+
+# Haystack configuration (default is minimal working configuration)
+HAYSTACK_CONNECTIONS = config(
+    "HAYSTACK_CONNECTIONS",
+    default={
+        "default": {"ENGINE": "courses.search_indexes.ConfigurableElasticSearchEngine"}
+    },
+    formatter=json.loads,
+)
+
+CKEDITOR_UPLOAD_PATH = "./"
+
+# ### FUN-APPS SETTINGS ###
+# -- Base --
+FUN_BASE_ROOT = path(os.path.dirname(imp.find_module("funsite")[1]))
+
+# Add 'theme/cms/templates' directory to MAKO template finder to override some
+# CMS templates
+MAKO_TEMPLATES["main"] = [FUN_BASE_ROOT / "fun/templates/cms"] + MAKO_TEMPLATES["main"]
+
+# Max size of asset uploads to GridFS
+MAX_ASSET_UPLOAD_FILE_SIZE_IN_MB = config(
+    "MAX_ASSET_UPLOAD_FILE_SIZE_IN_MB", default=10, formatter=int
+)
+
+# Force Edx to use `libcast_xblock` as default video player
+# in the studio (big green button) and if any xblock is called `video`
+XBLOCK_SELECT_FUNCTION = prefer_fun_video
